UX-434 🐙🔃 added more helper files for cleaner organization
