+++ packages/ListBox/test/helpers/toggleHelpers.js
@@ -1,0 +1,9 @@
+import selectors from "./selectors";
+
+export function toggleLazyDropDown() {
+  cy.contains("Marvel API").click();
+}
+
+export function toggleDropdown() {
+  cy.get(selectors.trigger).click();
+}
